+++ omnigibson/systems/macro_particle_system.py
@@ -597,7 +597,7 @@ class MacroVisualParticleSystem(MacroParticleSystem, VisualParticleSystem):
         obj = self._group_objects[group]
 
         # Sample scales and corresponding bbox extents
-        scales = self.sample_scales_by_group(group=group, n=max_samples).float()
+        scales = self.sample_scales_by_group(group=group, n=max_samples)
         # For sampling particle positions, we need the global bbox extents, NOT the local extents
         # which is what we would get naively if we directly use @scales
         avg_scale = th.pow(th.prod(obj.scale), 1 / 3)
